backend/pyproject.toml
[tool.poetry]
name = "stem-kit-backend"
version = "0.1.0"
description = "STEM Kit Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 